requirements.txt
aiogram>=3.0.0
google-genai>=0.2.0
google>=3.0.0
python-dotenv>=1.0.0
pymongo>=4.0.0
motor>=3.0.0
fastapi[standard]
uvicorn
aiohttp